requirements.txt
langchain
streamlit
python-dotenv
langchain-google-genai